Upgrade minion connector versions

+++ magda-scala-common/src/main/scala/au/csiro/data61/magda/model/Misc.scala
@@ -343,7 +343,7 @@ package misc {
       rights: Option[String] = None,
       accessURL: Option[String] = None,
       downloadURL: Option[String] = None,
-      byteSize: Option[Int] = None,
+      byteSize: Option[Long] = None,
       mediaType: Option[MediaType] = None,
       source: Option[DataSouce] = None,
       format: Option[String] = None
